Split tracing middleware into two
